refactor into discrete sub-pkgs
